requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
langgraph-cli[inmem]
numpy
pandas
uvicorn
langchain_google_genai 
